% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cardio.R
\name{model_cardio}
\alias{model_cardio}
\title{Heart Disease Probability Function}
\usage{
model_cardio(age, sex, thalach)
}
\arguments{
\item{age}{age of person as of today}

\item{sex}{Female or Male}

\item{thalach}{blood pressure}
}
\value{
The odds of having a hear disease
}
\description{
Heart Disease Probability Function
}
\examples{
model_cardio(30, "Female" , 120)
\dontrun{
model_cardio(52, "Male" , 120)
}
}
