% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_cardio.R
\name{model_cardio}
\alias{model_cardio}
\title{Model Cardio function}
\usage{
model_cardio(age, sex, thalach)
}
\arguments{
\item{age}{edad del paciente}

\item{sex}{género del paciente ("Female","Male")}

\item{thalach}{presión sanguínea del paciente}
}
\value{

}
\description{
Model Cardio function
}
\examples{
model_cardio(30,"Female",100)
}
